Time: 24 ms (33.60%) | Memory: 17.4 MB (57.86%) - LeetSync

+++ 18-4sum/4sum.cpp
@@ -3,12 +3,12 @@ public:
     vector<vector<int>> fourSum(vector<int>& nums, int target) {
         vector<vector<int>> ans;
         sort(nums.begin(), nums.end());
-        for(int i=0; i<nums.size(); i++){
+        for(long long i=0; i<nums.size(); i++){
             if(i>0 && nums[i]==nums[i-1]){continue; }
-            for(int j=i+1; j<nums.size(); j++){
+            for(long long j=i+1; j<nums.size(); j++){
             if(j>i+1 && nums[j]==nums[j-1]){continue;}
-            int start=j+1;
-            int end=nums.size()-1;
+            long long start=j+1;
+            long long end=nums.size()-1;
             while(start<end){
                 long long sum= (long long)nums[i]+nums[j]+nums[start]+nums[end];
                 if(sum==target){
